truffle: Replace truffle.js to truffle-config.js deploy: Add deploy EnvStorage
