+++ contracts.json
@@ -1,1 +1,1 @@
-{"REGISTRY_ADDRESS":"0x632af7ffefd015e942a268b583acb531e9f17e78","STAKING_ADDRESS":"0x5281d6fd5155af1defb973c0847a9ecaf8f2e6b0","BALLOT_STORAGE_ADDRESS":"0xcd96e3a842f989d2fd15cd8d3a29260a1dc0d789","GOV_IMP_ADDRESS":"0xa0e212be11c81e81a517240254174d2f08a5ccf4","GOV_ADDRESS":"0xe06c67cf4a697baf37868bdf96eb576b44794a1a","GOV_DELEGATOR":"0xe06c67cf4a697baf37868bdf96eb576b44794a1a"}
+{"REGISTRY_ADDRESS":"0x054db73e099dd6584ebec5a2b16c90fdc9be0284","STAKING_ADDRESS":"0xc87ccb3fab23a05f5f16164c3e83271d8da80b8a","ENV_STORAGE_ADDRESS":"0x096e5372483daad4abdc726be6b4a000e66e5a58","BALLOT_STORAGE_ADDRESS":"0x633cf9e75f50c69a6d13e101968800a4319a92d6","GOV_ADDRESS":"0x1d5bd524929cd0a6170bfe1d4515864129ad5b98"}
