+++ truffle-config.js
@@ -21,12 +21,11 @@ require('babel-register')({
   ignore: /node_modules\/(?!zeppelin-solidity)/,
 });
 require('babel-polyfill');
+var HDWalletProvider = require("truffle-hdwallet-provider")
 
-// const HDWalletProvider = require('truffle-hdwallet-provider');
-// const infuraKey = "fj4jll3k.....";
-//
-// const fs = require('fs');
-// const mnemonic = fs.readFileSync(".secret").toString().trim();
+const config = require('config')
+const ropstenConfig = config.get('ropsten')
+const metaTestnetConfig = config.get('metadiumTestnet')
 
 module.exports = {
   /**
@@ -47,65 +46,40 @@ module.exports = {
     // options below to some value.
     //
     development: {
-      host: '127.0.0.1', // Localhost (default: none)
-      port: 8545, // Standard Ethereum port (default: none)
-      network_id: '*', // Any network (default: none)
-      gas: 1000000000,
+      host: "localhost",
+      port: 8545,
+      network_id: "*", // Match any network id
+      //gas: 20000000000
     },
-
-    // Another network with more advanced options...
-    // advanced: {
-    // port: 8777,             // Custom port
-    // network_id: 1342,       // Custom network
-    // gas: 8500000,           // Gas sent with each transaction (default: ~6700000)
-    // gasPrice: 20000000000,  // 20 gwei (in wei) (default: 100 gwei)
-    // from: <address>,        // Account to send txs from (default: accounts[0])
-    // websockets: true        // Enable EventEmitter interface for web3 (default: false)
-    // },
-
-    // Useful for deploying to a public network.
-    // NB: It's important to wrap the provider as a function.
-    // ropsten: {
-    // provider: () => new HDWalletProvider(mnemonic, `https://ropsten.infura.io/${infuraKey}`),
-    // network_id: 3,       // Ropsten's id
-    // gas: 5500000,        // Ropsten has a lower block limit than mainnet
-    // confirmations: 2,    // # of confs to wait between deployments. (default: 0)
-    // timeoutBlocks: 200,  // # of blocks before a deployment times out  (minimum/default: 50)
-    // skipDryRun: true     // Skip dry run before migrations? (default: false for public nets )
-    // },
-
-    // Useful for private networks
-    // private: {
-    // provider: () => new HDWalletProvider(mnemonic, `https://network.io`),
-    // network_id: 2111,   // This network is yours, in the cloud.
-    // production: true    // Treats this network as if it was a public net. (default: false)
-    // }
-    coverage: {
-      host: 'localhost',
-      network_id: '*',
-      port: 8555, // <-- If you change this, also set the port option in .solcover.js.
-      gas: 0xfffffffffff, // <-- Use this high gas value
-      gasPrice: 0x01, // <-- Use this low gas price
+    metadiumTestnet: {
+        provider: () => {
+          return new HDWalletProvider(metaTestnetConfig.mnemonic, metaTestnetConfig.provider);
+        },
+        network_id: metaTestnetConfig.network_id,
+        gasPrice: metaTestnetConfig.gasPrice
     },
+    ropsten: {
+        provider: function() {
+          return new HDWalletProvider(ropstenConfig.mnemonic, ropstenConfig.provider)
+        },
+        network_id: ropstenConfig.network_id,
+        gas: ropstenConfig.gas,
+        gasPrice: ropstenConfig.gasPrice
+    }
   },
 
   // Set default mocha options here, use special reporters etc.
   mocha: {
-    // timeout: 100000
+
   },
 
   // Configure your compilers
   compilers: {
     solc: {
-      // version: "0.5.1",    // Fetch exact version from solc-bin (default: truffle's version)
-      // docker: true,        // Use "0.5.1" you've installed locally with docker (default: false)
-      // settings: {          // See the solidity docs for advice about optimization and evmVersion
-      //  optimizer: {
-      //    enabled: false,
-      //    runs: 200
-      //  },
-      //  evmVersion: "byzantium"
-      // }
+      optimizer: {
+        enabled: true,
+        runs: 200
+      },
     },
   },
 };
